Fixed small error, upgrade to ruby 3.3.0
I wanted to try this on 3.3.0 so I updated the ruby version again

I removed the `debug` gem since it was not part of the assignment,
and was only used so I could identify the bug in the original code.

I renamed `created_at` to `used_at` since it was getting updated now
to indicate the last used item.

The bug that was causing problems was not realizing that `reject` is
really a loop that was calling `evictee` multiple times. Assigning the
value of `evictee` to a local variable allowed me to make sure this was
only called once per call to `evict` and fix the bug.